requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==37.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.0.5
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-storages
django-allauth==0.51.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-storages[google]==1.12.3
    # via -r requirements.in
google-api-core[grpc]==2.8.2
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.9.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.1
    # via google-cloud-storage
google-cloud-secret-manager==2.11.1
    # via -r requirements.in
google-cloud-storage==2.4.0
    # via django-storages
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.56.3
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.4
    # via google-cloud-secret-manager
grpcio==1.47.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.47.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
oauthlib==3.2.0
    # via requests-oauthlib
proto-plus==1.20.6
    # via google-cloud-secret-manager
protobuf==3.20.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.4.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pyyaml==6.0
    # via -r requirements.in
requests==2.28.0
    # via
    #   django-allauth
    #   google-api-core
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   -r requirements.in
    #   google-auth
    #   grpcio
sqlparse==0.4.2
    # via django
urllib3==1.26.9
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
